requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
python-jose[cryptography]
python-multipart
requests
